visitors-managment/backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary
python-decouple==3.8
django-filter==23.3
Pillow
openpyxl==3.1.2
dj-database-url==2.1.0
gunicorn==21.2.0
python-docx==1.1.0
whitenoise==6.7.0